Move time utils to util/clock package
